refactor(service): include original_amount in OutgoingRequest
this allows us to decouple the exchange rate and balance services (before, they needed to be together because the balance service needs access to the incoming and outgoing amounts but only the exchange rate service had access to both)
this also makes it easier to implement an outgoing service that tracks packet statistics or saves them to some persistence layer, because such a service will also want access to the incoming and outgoing amounts
note that the field is called "original_amount" only because it seemed strange to add an "incoming_amount" for the services that create OutgoingRequests on their own (only CCP for now but this would also apply to a multi-path STREAM sender that directly creates OutgoingRequests)
