+++ crates/interledger-service-util/src/lib.rs
@@ -6,19 +6,19 @@
 extern crate log;
 
 mod echo;
+mod balances;
+mod exchange_rates;
 mod expiry_shortener;
 mod max_packet_amount;
 mod rate_limit;
-mod rates_and_balances;
 mod validator;
 
 pub use self::echo::EchoService;
+pub use self::balances::{BalanceService, BalanceStore};
+pub use self::exchange_rates::{ExchangeRateService, ExchangeRateStore};
 pub use self::expiry_shortener::{
     ExpiryShortenerService, RoundTripTimeAccount, DEFAULT_ROUND_TRIP_TIME,
 };
 pub use self::max_packet_amount::{MaxPacketAmountAccount, MaxPacketAmountService};
 pub use self::rate_limit::{RateLimitAccount, RateLimitError, RateLimitService, RateLimitStore};
-pub use self::rates_and_balances::{
-    BalanceStore, ExchangeRateAndBalanceService, ExchangeRateStore,
-};
 pub use self::validator::ValidatorService;
